Rename: 6주차미션 안으로 폴더 위치 변경
