+++ 해밍_황민경/6주차미션/미션/.gitignore
@@ -12,6 +12,7 @@ dist
 dist-ssr
 *.local
 .env
+mission_2/
 
 # Editor directories and files
 .vscode/*
