+++ 해밍_황민경/6주차미션/미션/src/components/Comment/CommentInput.tsx
